refactor : 서비스 부분 리팩토링 메모리 저장소 서비스와 db 서비스가 내용이 같아서 서비스를 하나로 사용
