+++ src/main/java/kr/codesqaud/cafe/controller/ArticleController.java
@@ -2,40 +2,30 @@ package kr.codesqaud.cafe.controller;
 
 import kr.codesqaud.cafe.domain.article.Article;
 import kr.codesqaud.cafe.dto.ArticleFormDto;
-import kr.codesqaud.cafe.service.jdbc.ArticleJdbcService;
-import kr.codesqaud.cafe.service.memory.ArticleMemoryService;
-import org.springframework.beans.factory.annotation.Autowired;
+import kr.codesqaud.cafe.service.ArticleService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
-import java.util.List;
 
 @Controller
 public class ArticleController {
-
-    private  ArticleMemoryService articleMemoryService;
-
-    public ArticleController(ArticleMemoryService articleMemoryService) {
-        this.articleMemoryService = articleMemoryService;
-    }
-    private ArticleJdbcService articleJdbcService;
-    @Autowired
-    public ArticleController(ArticleJdbcService articleJdbcService) {
-        this.articleJdbcService = articleJdbcService;
+    private final ArticleService articleService;
+    public ArticleController(ArticleService articleService) {
+        this.articleService = articleService;
     }
 
     @PostMapping("/article/write")
     public String postQuestion(ArticleFormDto articleFormDto) {
-        articleJdbcService.writeArticle(articleFormDto);
+        articleService.writeArticle(articleFormDto);
         return "redirect:/";
     }
 
     @GetMapping("/article/show/{index}")
     public String getShow(@PathVariable int index, Model model) {
-        Article article = articleJdbcService.findByIDX(index);
+        Article article = articleService.findByIDX(index);
         model.addAttribute("article", article);
         return "qna/show";
     }
@@ -43,7 +33,7 @@ public class ArticleController {
 
     @GetMapping("/")
     public String getIndex(Model model) {
-        model.addAttribute("articleList", articleJdbcService.getAricleList());
+        model.addAttribute("articleList", articleService.getAricleList());
         return "index";
     }
