+++ src/main/java/kr/codesqaud/cafe/controller/UserController.java
@@ -3,9 +3,7 @@ package kr.codesqaud.cafe.controller;
 import kr.codesqaud.cafe.dto.SignUpFormDto;
 import kr.codesqaud.cafe.dto.UpdateFormDto;
 import kr.codesqaud.cafe.domain.user.User;
-import kr.codesqaud.cafe.service.jdbc.UserJdbcService;
-import kr.codesqaud.cafe.service.memory.UserMemoryService;
-import org.springframework.beans.factory.annotation.Autowired;
+import kr.codesqaud.cafe.service.UserService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -15,36 +13,29 @@ import java.util.List;
 @Controller
 @RequestMapping("/user")
 public class UserController {
-    private UserMemoryService userMemoryService;
+    private final UserService userService;
 
-    public UserController(UserMemoryService userMemoryService) {
-        this.userMemoryService = userMemoryService;
-    }
-
-    private UserJdbcService userJdbcService;
-
-    @Autowired
-    public UserController(UserJdbcService userJdbcService) {
-        this.userJdbcService = userJdbcService;
+    public UserController(UserService userService) {
+        this.userService = userService;
     }
 
 
     @PostMapping("/create")
     public String postSignUp(SignUpFormDto signUpFormDto) {
-        userJdbcService.signUp(signUpFormDto);
+        userService.signUp(signUpFormDto);
         return "redirect:/user";
     }
 
     @GetMapping("")
     public String getUserList(Model model) {
-        List<User> list = userJdbcService.users();
+        List<User> list = userService.users();
         model.addAttribute("memberList", list);
         return "user/list";
     }
 
     @GetMapping("/profile/{id}")
     public String getProfile(@PathVariable String id, Model model) {
-        User user = userJdbcService.findById(id);
+        User user = userService.findById(id);
         model.addAttribute("name", user.getName());
         model.addAttribute("email", user.getEmail());
         return "user/profile";
@@ -52,14 +43,14 @@ public class UserController {
 
     @GetMapping("/update/{id}")
     public String getUpdateForm(@PathVariable String id, Model model) {
-        User user = userJdbcService.findById(id);
+        User user = userService.findById(id);
         model.addAttribute("user", user);
         return "user/update_form";
     }
 
     @PutMapping("/update/{id}")
     public String putUpdate(@PathVariable String id, UpdateFormDto updateFormDto) {
-        userJdbcService.update(userJdbcService.findById(id), updateFormDto);
+        userService.update(userService.findById(id), updateFormDto);
         return "redirect:/user";
     }
