+++ src/main/java/kr/codesqaud/cafe/service/UserService.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.service.jdbc;
+package kr.codesqaud.cafe.service;
 
 import kr.codesqaud.cafe.domain.user.User;
 import kr.codesqaud.cafe.domain.user.repository.UserRepository;
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class UserJdbcService {
+public class UserService {
     private final UserRepository userRepository;
 
-    public UserJdbcService(UserRepository userRepository) {
+    public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
