+++ src/main/resources/templates/layout/header.html
@@ -31,7 +31,7 @@
                         <li><a href="https://facebook.com" target="_blank">Facebook</a></li>
                     </ul>
                 </li>
-                <li><a href="../user/list.html"><i class="glyphicon glyphicon-user"></i></a></li>
+                <li><a href="/user"><i class="glyphicon glyphicon-user"></i></a></li>
             </ul>
         </div>
     </div>
